FU-212 feat: 의뢰자측 신청서 취소 API 개발 및 신청서 관련 공통로직 추출
사진작가 측 신청서 상태변경 메소드를 의뢰자측 신청서 취소 요청 시에 공통으로 사용하고자 ReservationService로 분리하였음.
