+++ src/main/java/com/foru/freebe/reservation/service/PhotographerReservationDetails.java
@@ -7,35 +7,26 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 
 import com.foru.freebe.common.dto.ApiResponse;
-import com.foru.freebe.errors.errorcode.CommonErrorCode;
-import com.foru.freebe.errors.exception.RestApiException;
 import com.foru.freebe.reservation.dto.CustomerDetails;
 import com.foru.freebe.reservation.dto.FormDetailsViewResponse;
 import com.foru.freebe.reservation.dto.PreferredDate;
 import com.foru.freebe.reservation.dto.ReferenceImageUrls;
-import com.foru.freebe.reservation.dto.ReservationStatusUpdateRequest;
 import com.foru.freebe.reservation.dto.StatusHistory;
 import com.foru.freebe.reservation.entity.ReferenceImage;
 import com.foru.freebe.reservation.entity.ReservationForm;
-import com.foru.freebe.reservation.entity.ReservationHistory;
 import com.foru.freebe.reservation.repository.ReferenceImageRepository;
-import com.foru.freebe.reservation.repository.ReservationFormRepository;
-import com.foru.freebe.reservation.repository.ReservationHistoryRepository;
 
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class PhotographerReservationDetails {
-    private final ReservationFormRepository reservationFormRepository;
-    private final ReservationHistoryRepository reservationHistoryRepository;
+    private final ReservationService reservationService;
     private final ReferenceImageRepository referenceImageRepository;
-    private final ReservationValidator reservationValidator;
 
     public ApiResponse<FormDetailsViewResponse> getReservationFormDetails(Long photographerId, Long formId) {
-        ReservationForm reservationForm = findReservationForm(photographerId, formId);
-        List<StatusHistory> statusHistories = getStatusHistories(reservationForm);
+        ReservationForm reservationForm = reservationService.findReservationForm(photographerId, formId, true);
+        List<StatusHistory> statusHistories = reservationService.getStatusHistories(reservationForm);
 
         CustomerDetails customerDetails = buildCustomerDetails(reservationForm);
         Map<String, String> shootDetails = reservationForm.getPhotoInfo();
@@ -56,55 +47,6 @@ public class PhotographerReservationDetails {
             .message("Successfully get reservation list")
             .status(200)
             .data(formDetailsViewResponse)
-            .build();
-    }
-
-    @Transactional
-    public ApiResponse<Void> updateReservationStatus(Long photographerId, Long formId,
-        ReservationStatusUpdateRequest request) {
-
-        ReservationForm reservationForm = findReservationForm(photographerId, formId);
-        reservationValidator.validateStatusChange(reservationForm.getReservationStatus(), request.getUpdateStatus());
-
-        reservationForm.updateReservationStatus(request.getUpdateStatus());
-        ReservationHistory history = getReservationHistory(request, reservationForm);
-
-        reservationFormRepository.save(reservationForm);
-        reservationHistoryRepository.save(history);
-
-        return ApiResponse.<Void>builder()
-            .message("Successfully update reservation status")
-            .status(200)
-            .build();
-    }
-
-    private ReservationHistory getReservationHistory(ReservationStatusUpdateRequest request,
-        ReservationForm reservationForm) {
-        if (request.getCancellationReason() != null) {
-            return ReservationHistory.createCancelledReservationHistory(reservationForm, request.getUpdateStatus(),
-                request.getCancellationReason());
-        } else {
-            return ReservationHistory.createReservationHistory(reservationForm, request.getUpdateStatus());
-        }
-    }
-
-    private ReservationForm findReservationForm(Long photographerId, Long formId) {
-        return reservationFormRepository.findByPhotographerIdAndId(photographerId, formId)
-            .orElseThrow(() -> new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND));
-    }
-
-    private List<StatusHistory> getStatusHistories(ReservationForm reservationForm) {
-        return reservationHistoryRepository.findAllByReservationForm(reservationForm)
-            .orElseThrow(() -> new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND))
-            .stream()
-            .map(this::toStatusHistory)
-            .collect(Collectors.toList());
-    }
-
-    private StatusHistory toStatusHistory(ReservationHistory reservationHistory) {
-        return StatusHistory.builder()
-            .reservationStatus(reservationHistory.getReservationStatus())
-            .statusUpdateDate(reservationHistory.getStatusUpdateDate().toLocalDate())
             .build();
     }
